+++ sdk/trace/trace_test.go
@@ -26,17 +26,16 @@ import (
 	"time"
 
 	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/label"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	grpccodes "google.golang.org/grpc/codes"
 
 	"go.opentelemetry.io/otel/api/apitest"
 	"go.opentelemetry.io/otel/api/trace"
 	apitrace "go.opentelemetry.io/otel/api/trace"
-	otelcodes "go.opentelemetry.io/otel/codes"
 	ottest "go.opentelemetry.io/otel/internal/testing"
 	export "go.opentelemetry.io/otel/sdk/export/trace"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
@@ -649,7 +648,7 @@ func TestSetSpanStatus(t *testing.T) {
 	tp := NewTracerProvider(WithSyncer(te))
 
 	span := startSpan(tp, "SpanStatus")
-	span.SetStatus(otelcodes.Canceled, "canceled")
+	span.SetStatus(codes.Error, "Error")
 	got, err := endSpan(te, span)
 	if err != nil {
 		t.Fatal(err)
@@ -663,8 +662,8 @@ func TestSetSpanStatus(t *testing.T) {
 		ParentSpanID:           sid,
 		Name:                   "span0",
 		SpanKind:               apitrace.SpanKindInternal,
-		StatusCode:             grpccodes.Canceled,
-		StatusMessage:          "canceled",
+		StatusCode:             codes.Error,
+		StatusMessage:          "Error",
 		HasRemoteParent:        true,
 		InstrumentationLibrary: instrumentation.Library{Name: "SpanStatus"},
 	}
@@ -1027,7 +1026,7 @@ func TestRecordErrorWithStatus(t *testing.T) {
 
 	testErr := ottest.NewTestError("test error")
 	errTime := time.Now()
-	testStatus := otelcodes.Unknown
+	testStatus := codes.Error
 	span.RecordError(context.Background(), testErr,
 		apitrace.WithErrorTime(errTime),
 		apitrace.WithErrorStatus(testStatus),
@@ -1046,7 +1045,7 @@ func TestRecordErrorWithStatus(t *testing.T) {
 		ParentSpanID:    sid,
 		Name:            "span0",
 		SpanKind:        apitrace.SpanKindInternal,
-		StatusCode:      grpccodes.Unknown,
+		StatusCode:      codes.Error,
 		StatusMessage:   "",
 		HasRemoteParent: true,
 		MessageEvents: []export.Event{
@@ -1087,7 +1086,7 @@ func TestRecordErrorNil(t *testing.T) {
 		Name:                   "span0",
 		SpanKind:               apitrace.SpanKindInternal,
 		HasRemoteParent:        true,
-		StatusCode:             grpccodes.OK,
+		StatusCode:             codes.Unset,
 		StatusMessage:          "",
 		InstrumentationLibrary: instrumentation.Library{Name: "RecordErrorNil"},
 	}